requirements.txt
requests
openai
Pillow
moviepy
python-dotenv
tweepy
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
